pyproject.toml
[tool.poetry]
name = "transcription-battle"
version = "0.1.0"
description = ""
authors = ["Luiz Pereira <luiz@fireworkhq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyaudio = "^0.2.14"
google-cloud-speech = "^2.25.0"
deepgram-sdk = "^3.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
